제거까지 오류 수정
장바구니 요소 제거까지 완료
비었을때 비었다는 페이지, 주문하기 버튼 비활성화 추가

+++ src/app/data/checkbox-value.ts
@@ -110,12 +110,15 @@ export class CheckboxValue extends CheckValue{
         this.items = items;
     }
 
-    deleteItem(item: CheckValue){
-        this.parent.items.splice(this.parent.items.indexOf(item), 1);
-        this.parent.checkEmpty();
+    deleteItem(child: CheckValue){
+        this.items.splice(this.items.indexOf(child), 1);
+        this.checkEmpty();
     }
 
     checkEmpty(){
+        if(this.parent == null){
+            return;
+        }
         if(this.items.length == 0){
             this.parent.deleteItem(this);
         }
